@@ -10,7 +10,6 @@ use OpenApi\Context;
 use OpenApi\Generator;
 use OpenApi\Annotations as OA;
 use OpenApi\OpenApiException;
-use OpenApi\Util;
 use Symfony\Component\Yaml\Yaml;
 
 /**
@@ -141,30 +140,6 @@ abstract class AbstractAnnotation implements \JsonSerializable
                 }
             }
         }
-
-        if ($this instanceof OpenApi) {
-            if ($this->_context->root()->version) {
-                // override via `Generator::setVersion()`
-                $this->openapi = $this->_context->root()->version;
-            } else {
-                $this->_context->root()->version = $this->openapi;
-            }
-        }
-    }
-
-    public function __get(string $property)
-    {
-        $properties = get_object_vars($this);
-        $this->_context->logger->warning('Property "' . $property . '" doesn\'t exist in a ' . $this->identity() . ', existing properties: "' . implode('", "', array_keys($properties)) . '" in ' . $this->_context);
-    }
-
-    public function __set(string $property, $value): void
-    {
-        $fields = get_object_vars($this);
-        foreach (static::$_blacklist as $_property) {
-            unset($fields[$_property]);
-        }
-        $this->_context->logger->warning('Ignoring unexpected property "' . $property . '" for ' . $this->identity() . ', expecting "' . implode('", "', array_keys($fields)) . '" in ' . $this->_context);
     }
 
     /**
@@ -239,7 +214,7 @@ abstract class AbstractAnnotation implements \JsonSerializable
                 if (Generator::isDefault($value)) {
                     continue;
                 }
-                $identity = method_exists($object, 'identity') ? $object->identity() : get_class($object);
+                $identity = method_exists($object, 'identity') ? $object->identity() : $object::class;
                 $context1 = $this->_context;
                 $context2 = property_exists($object, '_context') ? $object->_context : 'unknown';
                 if ($this->{$property} instanceof AbstractAnnotation) {
@@ -391,10 +366,17 @@ abstract class AbstractAnnotation implements \JsonSerializable
             if (isset($data->type) && is_array($data->type)) {
                 if (in_array('null', $data->type)) {
                     $data->nullable = true;
-                    $data->type = array_filter($data->type, fn ($t): bool => 'null' !== $t);
+                    $data->type = array_filter($data->type, static fn ($v): bool => $v !== 'null');
                     if (1 === count($data->type)) {
                         $data->type = array_pop($data->type);
                     }
+                }
+            }
+            if (isset($data->type) && is_array($data->type)) {
+                if (1 === count($data->type)) {
+                    $data->type = array_pop($data->type);
+                } else {
+                    unset($data->type);
                 }
             }
         }
@@ -461,19 +443,18 @@ abstract class AbstractAnnotation implements \JsonSerializable
                 break;
             }
 
-            /** @var class-string<AbstractAnnotation> $class */
-            $class = get_class($annotation);
+            $class = $annotation::class;
             if ($details = $this->matchNested($annotation)) {
                 $property = $details->value;
                 if (is_array($property)) {
-                    $this->_context->logger->warning('Only one ' . Util::shorten(get_class($annotation)) . '() allowed for ' . $this->identity() . ' multiple found, skipped: ' . $annotation->_context);
+                    $this->_context->logger->warning('Only one ' . static::shorten($annotation::class) . '() allowed for ' . $this->identity() . ' multiple found, skipped: ' . $annotation->_context);
                 } else {
-                    $this->_context->logger->warning('Only one ' . Util::shorten(get_class($annotation)) . '() allowed for ' . $this->identity() . " multiple found in:\n    Using: " . $this->{$property}->_context . "\n  Skipped: " . $annotation->_context);
+                    $this->_context->logger->warning('Only one ' . static::shorten($annotation::class) . '() allowed for ' . $this->identity() . " multiple found in:\n    Using: " . $this->{$property}->_context . "\n  Skipped: " . $annotation->_context);
                 }
             } elseif ($annotation instanceof AbstractAnnotation) {
                 $message = 'Unexpected ' . $annotation->identity();
                 if ($class::$_parents) {
-                    $message .= ', expected to be inside ' . implode(', ', Util::shorten($class::$_parents));
+                    $message .= ', expected to be inside ' . implode(', ', static::shorten($class::$_parents));
                 }
                 $this->_context->logger->warning($message . ' in ' . $annotation->_context);
             }
@@ -493,7 +474,7 @@ abstract class AbstractAnnotation implements \JsonSerializable
             $keyField = $nested[1];
             foreach ($this->{$property} as $key => $item) {
                 if (is_array($item) && is_numeric($key) === false) {
-                    $this->_context->logger->warning($this->identity() . '->' . $property . ' is an object literal, use nested ' . Util::shorten($annotationClass) . '() annotation(s) in ' . $this->_context);
+                    $this->_context->logger->warning($this->identity() . '->' . $property . ' is an object literal, use nested ' . static::shorten($annotationClass) . '() annotation(s) in ' . $this->_context);
                     $keys[$key] = $item;
                 } elseif (Generator::isDefault($item->{$keyField})) {
                     $this->_context->logger->error($item->identity() . ' is missing key-field: "' . $keyField . '" in ' . $item->_context);
@@ -506,7 +487,7 @@ abstract class AbstractAnnotation implements \JsonSerializable
         }
 
         if (property_exists($this, 'ref') && !Generator::isDefault($this->ref) && is_string($this->ref)) {
-            if (substr($this->ref, 0, 2) === '#/' && $stack !== [] && $stack[0] instanceof OpenApi) {
+            if (str_starts_with($this->ref, '#/') && $stack !== [] && $stack[0] instanceof OpenApi) {
                 // Internal reference
                 try {
                     $stack[0]->ref($this->ref);
@@ -523,11 +504,11 @@ abstract class AbstractAnnotation implements \JsonSerializable
                         $nestedProperty = is_array($nested) ? $nested[0] : $nested;
                         if ($property === $nestedProperty) {
                             if ($this instanceof OpenApi) {
-                                $message = 'Required ' . Util::shorten($class) . '() not found';
+                                $message = 'Required ' . static::shorten($class) . '() not found';
                             } elseif (is_array($nested)) {
-                                $message = $this->identity() . ' requires at least one ' . Util::shorten($class) . '() in ' . $this->_context;
+                                $message = $this->identity() . ' requires at least one ' . static::shorten($class) . '() in ' . $this->_context;
                             } else {
-                                $message = $this->identity() . ' requires a ' . Util::shorten($class) . '() in ' . $this->_context;
+                                $message = $this->identity() . ' requires a ' . static::shorten($class) . '() in ' . $this->_context;
                             }
                             break;
                         }
@@ -553,7 +534,7 @@ abstract class AbstractAnnotation implements \JsonSerializable
                     $this->_context->logger->warning($this->identity() . '->' . $property . ' "' . $value . '" is invalid, expecting "' . implode('", "', $type) . '" in ' . $this->_context);
                 }
             } else {
-                throw new OpenApiException('Invalid ' . get_class($this) . '::$_types[' . $property . ']');
+                throw new OpenApiException('Invalid ' . static::class . '::$_types[' . $property . ']');
             }
         }
         $stack[] = $this;
@@ -612,7 +593,7 @@ abstract class AbstractAnnotation implements \JsonSerializable
      */
     public function identity(): string
     {
-        $class = get_class($this);
+        $class = static::class;
         $properties = [];
         /** @var class-string<AbstractAnnotation> $parent */
         foreach (static::$_parents as $parent) {
@@ -653,10 +634,10 @@ abstract class AbstractAnnotation implements \JsonSerializable
      */
     public function getRoot(): string
     {
-        $class = get_class($this);
+        $class = static::class;
 
         do {
-            if (0 === strpos($class, 'OpenApi\\Annotations\\')) {
+            if (str_starts_with($class, 'OpenApi\\Annotations\\')) {
                 break;
             }
         } while ($class = get_parent_class($class));
@@ -671,7 +652,7 @@ abstract class AbstractAnnotation implements \JsonSerializable
      */
     public function isRoot(string $rootClass): bool
     {
-        return get_class($this) === $rootClass || $this->getRoot() === $rootClass;
+        return static::class === $rootClass || $this->getRoot() === $rootClass;
     }
 
     /**
@@ -687,7 +668,7 @@ abstract class AbstractAnnotation implements \JsonSerializable
             }
         }
 
-        return Util::shorten(get_class($this)) . '(' . implode(',', $fields) . ')';
+        return static::shorten(static::class) . '(' . implode(',', $fields) . ')';
     }
 
     /**
@@ -696,9 +677,9 @@ abstract class AbstractAnnotation implements \JsonSerializable
      * @param string $type  The annotations property type
      * @param mixed  $value The property value
      */
-    private function validateType(string $type, $value): bool
+    private function validateType(string $type, mixed $value): bool
     {
-        if (substr($type, 0, 1) === '[' && substr($type, -1) === ']') { // Array of a specified type?
+        if (str_starts_with($type, '[') && str_ends_with($type, ']')) { // Array of a specified type?
             if ($this->validateType('array', $value) === false) {
                 return false;
             }
@@ -725,7 +706,7 @@ abstract class AbstractAnnotation implements \JsonSerializable
      * @param string $type  The property type
      * @param mixed  $value The value to validate
      */
-    private function validateDefaultTypes(string $type, $value): bool
+    private function validateDefaultTypes(string $type, mixed $value): bool
     {
         if (str_contains($type, '|')) {
             $types = explode('|', $type);
@@ -738,24 +719,16 @@ abstract class AbstractAnnotation implements \JsonSerializable
             return false;
         }
 
-        switch ($type) {
-            case 'string':
-                return is_string($value);
-            case 'boolean':
-                return is_bool($value);
-            case 'integer':
-                return is_int($value);
-            case 'number':
-                return is_numeric($value);
-            case 'object':
-                return is_object($value);
-            case 'array':
-                return $this->validateArrayType($value);
-            case 'scheme':
-                return in_array($value, ['http', 'https', 'ws', 'wss'], true);
-            default:
-                throw new OpenApiException('Invalid type "' . $type . '"');
-        }
+        return match ($type) {
+            'string' => is_string($value),
+            'boolean' => is_bool($value),
+            'integer' => is_int($value),
+            'number' => is_numeric($value),
+            'object' => is_object($value),
+            'array' => $this->validateArrayType($value),
+            'scheme' => in_array($value, ['http', 'https', 'ws', 'wss'], true),
+            default => throw new OpenApiException('Invalid type "' . $type . '"'),
+        };
     }
 
     /**
@@ -772,7 +745,7 @@ abstract class AbstractAnnotation implements \JsonSerializable
             if ($count !== $i) {
                 return false;
             }
-            $count++;
+            ++$count;
         }
 
         return true;
@@ -780,11 +753,8 @@ abstract class AbstractAnnotation implements \JsonSerializable
 
     /**
      * Wrap the context with a reference to the annotation it is nested in.
-     *
-     *
-     * @return AbstractAnnotation
      */
-    protected function nested(AbstractAnnotation $annotation, Context $nestedContext)
+    protected function nested(AbstractAnnotation $annotation, Context $nestedContext): self
     {
         if (property_exists($annotation, '_context') && $annotation->_context === $this->_context) {
             $annotation->_context = $nestedContext;
@@ -804,6 +774,26 @@ abstract class AbstractAnnotation implements \JsonSerializable
             }
         }
 
-        return array_filter($combined, fn ($value): bool => !Generator::isDefault($value) && $value !== null);
+        return array_filter($combined, static fn ($value): bool => !Generator::isDefault($value) && $value !== null);
+    }
+
+    /**
+     * Shorten class name(s).
+     *
+     * @param array|object|string $classes Class(es) to shorten
+     *
+     * @return string|string[] One or more shortened class names
+     */
+    protected static function shorten($classes)
+    {
+        $short = [];
+        foreach ((array) $classes as $class) {
+            $short[] = '@' . str_replace([
+                    'OpenApi\\Annotations\\',
+                    'OpenApi\\Attributes\\',
+                ], 'OA\\', (string) $class);
+        }
+
+        return is_array($classes) ? $short : array_pop($short);
     }
 }
