@@ -8,28 +8,27 @@ namespace OpenApi\Tests\Analysers;
 
 use OpenApi\Analysers\TokenScanner;
 use OpenApi\Tests\OpenApiTestCase;
+use PHPUnit\Framework\Attributes\DataProvider;
 
 class TokenScannerTest extends OpenApiTestCase
 {
     public static function scanCases(): iterable
     {
-        if (\PHP_VERSION_ID >= 80100) {
-            yield 'abstract' => [
-                'PHP/AbstractKeyword.php',
-                [
-                    'OpenApi\Tests\Fixtures\PHP\AbstractKeyword' => [
-                        'uses' => [
-                            'OAT' => 'OpenApi\Attributes',
-                        ],
-                        'interfaces' => [],
-                        'traits' => [],
-                        'enums' => [],
-                        'methods' => ['stuff', 'other', 'another'],
-                        'properties' => [],
+        yield 'abstract' => [
+            'PHP/AbstractKeyword.php',
+            [
+                'OpenApi\Tests\Fixtures\PHP\AbstractKeyword' => [
+                    'uses' => [
+                        'OAT' => 'OpenApi\Attributes',
                     ],
+                    'interfaces' => [],
+                    'traits' => [],
+                    'enums' => [],
+                    'methods' => ['stuff', 'other', 'another'],
+                    'properties' => [],
                 ],
-            ];
-        }
+            ],
+        ];
 
         yield 'references' => [
             'PHP/References.php',
@@ -355,44 +354,40 @@ class TokenScannerTest extends OpenApiTestCase
             ],
         ];
 
-        if (\PHP_VERSION_ID >= 80100) {
-            yield 'enum' => [
-                'PHP/Enums/StatusEnum.php',
-                [
-                    'OpenApi\\Tests\\Fixtures\\PHP\\Enums\\StatusEnum' => [
-                        'uses' => [
-                            'OAT' => 'OpenApi\\Attributes',
-                        ],
-                        'interfaces' => [],
-                        'enums' => [],
-                        'traits' => [],
-                        'methods' => [],
-                        'properties' => [],
+        yield 'enum' => [
+            'PHP/Enums/StatusEnum.php',
+            [
+                'OpenApi\\Tests\\Fixtures\\PHP\\Enums\\StatusEnum' => [
+                    'uses' => [
+                        'OAT' => 'OpenApi\\Attributes',
                     ],
+                    'interfaces' => [],
+                    'enums' => [],
+                    'traits' => [],
+                    'methods' => [],
+                    'properties' => [],
                 ],
-            ];
+            ],
+        ];
 
-            yield 'enum-backed' => [
-                'PHP/Enums/StatusEnumBacked.php',
-                [
-                    'OpenApi\\Tests\\Fixtures\\PHP\\Enums\\StatusEnumBacked' => [
-                        'uses' => [
-                            'OAT' => 'OpenApi\\Attributes',
-                        ],
-                        'interfaces' => [],
-                        'enums' => [],
-                        'traits' => [],
-                        'methods' => [],
-                        'properties' => [],
+        yield 'enum-backed' => [
+            'PHP/Enums/StatusEnumBacked.php',
+            [
+                'OpenApi\\Tests\\Fixtures\\PHP\\Enums\\StatusEnumBacked' => [
+                    'uses' => [
+                        'OAT' => 'OpenApi\\Attributes',
                     ],
+                    'interfaces' => [],
+                    'enums' => [],
+                    'traits' => [],
+                    'methods' => [],
+                    'properties' => [],
                 ],
-            ];
-        }
+            ],
+        ];
     }
 
-    /**
-     * @dataProvider scanCases
-     */
+    #[DataProvider('scanCases')]
     public function testScanFile(string $fixture, array $expected): void
     {
         $result = (new TokenScanner())->scanFile($this->fixture($fixture));
