@@ -16,6 +16,7 @@ use OpenApi\Processors\Concerns\DocblockTrait;
 use OpenApi\Processors\MergeIntoComponents;
 use OpenApi\Processors\MergeIntoOpenApi;
 use OpenApi\Tests\OpenApiTestCase;
+use PHPUnit\Framework\Attributes\DataProvider;
 
 class AugmentParametersTest extends OpenApiTestCase
 {
@@ -54,9 +55,7 @@ class AugmentParametersTest extends OpenApiTestCase
         ];
     }
 
-    /**
-     * @dataProvider tagCases
-     */
+    #[DataProvider('tagCases')]
     public function testParseTags(string $params, array $expected): void
     {
         $mixed = $this->getContext(['comment' => "/**\n$params\n  *"]);
@@ -65,13 +64,11 @@ class AugmentParametersTest extends OpenApiTestCase
         $this->assertEquals($expected, $tags);
     }
 
-    /**
-     * @requires PHP 8.1
-     */
     public function testParameterNativeType(): void
     {
-        $analysis = $this->analysisFromFixtures(['RequestUsingAttribute.php']);
-        $analysis->process($this->initializeProcessors([
+        $analysis = $this->analysisFromFixtures([
+            'RequestUsingAttribute.php',
+        ], $this->processorPipeline([
             new MergeIntoOpenApi(),
             new MergeIntoComponents(),
             new BuildPaths(),
