@@ -1,5 +1,6 @@
 <?php
 
+use Rector\CodeQuality\Rector\Class_\CompleteDynamicPropertiesRector;
 use Rector\CodeQuality\Rector\For_\ForRepeatedCountToOwnVariableRector;
 use Rector\CodeQuality\Rector\If_\CombineIfRector;
 use Rector\CodeQuality\Rector\If_\ExplicitBoolCompareRector;
@@ -12,6 +13,7 @@ use Rector\Config\RectorConfig;
 use Rector\DeadCode\Rector\If_\RemoveAlwaysTrueIfConditionRector;
 use Rector\DeadCode\Rector\If_\RemoveDeadInstanceOfRector;
 use Rector\Php55\Rector\String_\StringClassNameToClassConstantRector;
+use Rector\Php80\Rector\Class_\ClassPropertyAssignToConstructorPromotionRector;
 use Rector\TypeDeclaration\Rector\ClassMethod\ParamTypeByMethodCallTypeRector;
 use Rector\ValueObject\PhpVersion;
 
@@ -40,7 +42,16 @@ return RectorConfig::configure()
         ParamTypeByMethodCallTypeRector::class => [
             __DIR__ . '/src/Serializer.php',
         ],
+        ClassPropertyAssignToConstructorPromotionRector::class,
+        CompleteDynamicPropertiesRector::class => [
+            __DIR__ . '/src/Annotations/AbstractAnnotation.php',
+        ],
     ])
-    ->withPreparedSets(true, true, true, true)
-    ->withPhpVersion(PhpVersion::PHP_74)
+    ->withPreparedSets(
+        deadCode: true,
+        codeQuality: true,
+        codingStyle: true,
+        typeDeclarations: true,
+    )
+    ->withPhpVersion(PhpVersion::PHP_82)
     ->withPhpSets();
