@@ -41,16 +41,16 @@ class DocBlockAnnotationFactory implements AnnotationFactoryInterface
         $aliases = $this->generator ? $this->generator->getAliases() : [];
 
         if (method_exists($reflector, 'getShortName') && method_exists($reflector, 'getName')) {
-            $aliases[strtolower($reflector->getShortName())] = $reflector->getName();
+            $aliases[strtolower((string) $reflector->getShortName())] = $reflector->getName();
         }
 
         if ($context->with('scanned')) {
             $details = $context->scanned;
             foreach ($details['uses'] as $alias => $name) {
-                $aliasKey = strtolower($alias);
+                $aliasKey = strtolower((string) $alias);
                 if ($name != $alias && !array_key_exists($aliasKey, $aliases)) {
                     // real aliases only
-                    $aliases[strtolower($alias)] = $name;
+                    $aliases[strtolower((string) $alias)] = $name;
                 }
             }
         }
