@@ -50,8 +50,7 @@ class Items extends Schema
         $valid = parent::validate($stack, $skip, $ref, $context);
 
         $parent = end($stack);
-        // type might be array in 3.1.0
-        if ($parent instanceof Schema && ($parent->type !== 'array' && !(is_array($parent->type) && in_array('array', $parent->type)))) {
+        if ($parent instanceof Schema && $parent->type !== 'array') {
             $this->_context->logger->warning('@OA\\Items() parent type must be "array" in ' . $this->_context);
             $valid = false;
         }
