@@ -32,8 +32,8 @@ use Psr\Log\LoggerInterface;
  * @property string|null                  $property
  * @property \Reflector|null              $reflector   Optional reflection details
  * @property bool|null                    $static      Indicate a static method
- * @property bool|null                    $generated   Indicate the context was generated by a processor or
- *                                                     the serializer
+ * @property bool|null                    $generated   Indicate the context was generated by a processor,
+ *                                                     type resolver or serializer
  * @property OA\AbstractAnnotation|null   $nested
  * @property OA\AbstractAnnotation[]|null $annotations
  * @property OA\AbstractAnnotation[]|null $other       Annotations not related to OpenApi
@@ -42,7 +42,7 @@ use Psr\Log\LoggerInterface;
  * @property string|null                  $version     The OpenAPI version in use
  */
 #[\AllowDynamicProperties]
-class Context
+class Context implements \Stringable
 {
     /**
      * Prototypical inheritance for properties.
@@ -183,7 +183,7 @@ class Context
 
     public function __serialize(): array
     {
-        return array_filter(get_object_vars($this), function ($value): bool {
+        return array_filter(get_object_vars($this), static function ($value): bool {
             $rc = is_object($value) ? new \ReflectionClass($value) : null;
 
             return (!$rc || !$rc->isAnonymous())
@@ -202,7 +202,7 @@ class Context
     /**
      * Traverse the context tree to get the property value.
      */
-    public function __get(string $property)
+    public function __get(string $property): mixed
     {
         if ($this->parent instanceof Context) {
             return $this->parent->{$property};
@@ -255,7 +255,7 @@ class Context
         } elseif ($this->uses) {
             // Unqualified name (Foo)
             foreach ($this->uses as $alias => $aliasedNamespace) {
-                if (strcasecmp($alias, $source) === 0) {
+                if (strcasecmp((string) $alias, $source) === 0) {
                     return '\\' . $aliasedNamespace;
                 }
             }
