@@ -65,14 +65,23 @@ class AugmentParameters implements GeneratorAwareInterface
             }
 
             if ($context->reflector instanceof \ReflectionParameter) {
-                $schema = new OA\Schema(['_context' => new Context(['reflector' => $context->reflector], $context)]);
+                $schema = new OA\Schema([
+                    '_context' => new Context([
+                        'generated' => true,
+                        'reflector' => $context->reflector,
+                    ], $context),
+                ]);
                 $this->generator->getTypeResolver()->augmentSchemaType($analysis, $schema);
 
                 $parameter->merge([new OA\Schema([
                     'type' => $schema->type,
                     'format' => $schema->format,
                     'ref' => $schema->ref,
-                    '_context' => new Context(['nested' => $this, 'comment' => null, 'reflector' => $context->reflector], $context)]),
+                    '_context' => new Context([
+                        'nested' => $this,
+                        'comment' => null,
+                        'reflector' => $context->reflector,
+                    ], $context)]),
                 ]);
 
                 if (Generator::isDefault($parameter->required)) {
@@ -113,7 +122,6 @@ class AugmentParameters implements GeneratorAwareInterface
 
     protected function augmentOperationParameters(Analysis $analysis): void
     {
-        /** @var OA\Operation[] $operations */
         $operations = $analysis->getAnnotationsOfType(OA\Operation::class);
 
         foreach ($operations as $operation) {
