@@ -24,7 +24,6 @@ class AugmentProperties implements GeneratorAwareInterface
 
     public function __invoke(Analysis $analysis): void
     {
-        /** @var OA\Property[] $properties */
         $properties = $analysis->getAnnotationsOfType(OA\Property::class);
 
         foreach ($properties as $property) {
@@ -57,7 +56,7 @@ class AugmentProperties implements GeneratorAwareInterface
                 $typeAndDescription = $this->parseVarLine((string) $context->comment);
 
                 if ($typeAndDescription['description']) {
-                    $property->description = $typeAndDescription['description'];
+                    $property->description = trim($typeAndDescription['description']);
                 } elseif ($this->isDocblockRoot($property)) {
                     $property->description = $this->parseDocblock($context->comment);
                 }
