@@ -20,7 +20,6 @@ class AugmentSchemas
 {
     public function __invoke(Analysis $analysis): void
     {
-        /** @var OA\Schema[] $schemas */
         $schemas = $analysis->getAnnotationsOfType(OA\Schema::class);
 
         $this->augmentSchema($schemas);
@@ -103,7 +102,7 @@ class AugmentSchemas
                     $schema->type = 'object';
                 }
             } else {
-                if (is_string($schema->type) && $typeSchema = $analysis->getSchemaForSource($schema->type)) {
+                if (is_string($schema->type) && $typeSchema = $analysis->getAnnotationForSource($schema->type)) {
                     if (Generator::isDefault($schema->format)) {
                         $schema->ref = OA\Components::ref($typeSchema);
                         $schema->type = Generator::UNDEFINED;
