@@ -19,7 +19,7 @@ trait AnnotationTrait
     {
         $storage = new \SplObjectStorage();
 
-        $this->traverseAnnotations($root, function ($item) use (&$storage): void {
+        $this->traverseAnnotations($root, static function ($item) use (&$storage): void {
             if ($item instanceof OA\AbstractAnnotation && !$storage->contains($item)) {
                 $storage->attach($item);
             }
@@ -34,7 +34,7 @@ trait AnnotationTrait
     public function removeAnnotation(iterable $root, OA\AbstractAnnotation $annotation, bool $recurse = true): void
     {
         $remove = $this->collectAnnotations($annotation);
-        $this->traverseAnnotations($root, function ($item) use ($remove): void {
+        $this->traverseAnnotations($root, static function ($item) use ($remove): void {
             if ($item instanceof \SplObjectStorage) {
                 foreach ($remove as $annotation) {
                     $item->detach($annotation);
