@@ -63,7 +63,6 @@ class ExpandEnums implements GeneratorAwareInterface
 
     protected function expandContextEnum(Analysis $analysis): void
     {
-        /** @var OA\Schema[] $schemas */
         $schemas = $analysis->getAnnotationsOfType(OA\Schema::class, true);
 
         foreach ($schemas as $schema) {
@@ -74,20 +73,17 @@ class ExpandEnums implements GeneratorAwareInterface
                 $schemaType = $schema->type;
                 $enumType = null;
                 if ($re->isBacked()) {
-                    $backingType = $re->getBackingType();
-                    if ($backingType instanceof \ReflectionNamedType) {
-                        $enumType = $backingType->getName();
-                    }
+                    $enumType = $re->getBackingType()->getName();
                 }
 
                 // no (or invalid) schema type means name
                 $useName = Generator::isDefault($schemaType) || ($enumType && $this->generator->getTypeResolver()->native2spec($enumType) != $schemaType);
 
-                $schema->enum = array_map(fn (\ReflectionEnumUnitCase $case) => ($useName || !($case instanceof \ReflectionEnumBackedCase)) ? $case->name : $case->getBackingValue(), $re->getCases());
+                $schema->enum = array_map(static fn (\ReflectionEnumUnitCase $case): int|string => ($useName || !($case instanceof \ReflectionEnumBackedCase)) ? $case->name : $case->getBackingValue(), $re->getCases());
 
                 if ($this->enumNames !== null && !$useName) {
                     $schemaX = Generator::isDefault($schema->x) ? [] : $schema->x;
-                    $schemaX[$this->enumNames] = array_map(fn (\ReflectionEnumUnitCase $case): string => $case->name, $re->getCases());
+                    $schemaX[$this->enumNames] = array_map(static fn (\ReflectionEnumUnitCase $case): string => $case->name, $re->getCases());
 
                     $schema->x = $schemaX;
                 }
@@ -101,7 +97,6 @@ class ExpandEnums implements GeneratorAwareInterface
 
     protected function expandSchemaEnum(Analysis $analysis): void
     {
-        /** @var OA\Schema[] $schemas */
         $schemas = $analysis->getAnnotationsOfType([OA\Schema::class, OA\ServerVariable::class]);
 
         foreach ($schemas as $schema) {
