@@ -8,7 +8,6 @@ namespace OpenApi\Type;
 
 use OpenApi\Analysis;
 use OpenApi\Annotations as OA;
-use OpenApi\Context;
 use OpenApi\Generator;
 use OpenApi\TypeResolverInterface;
 
@@ -16,40 +15,12 @@ abstract class AbstractTypeResolver implements TypeResolverInterface
 {
     protected function type2ref(OA\Schema $schema, Analysis $analysis, string $sourceClass = OA\Schema::class): void
     {
-        if (!Generator::isDefault($schema->type)) {
+        if (!Generator::isDefault($schema->type) && !is_array($schema->type)) {
             if ($typeSchema = $analysis->getAnnotationForSource($schema->type, $sourceClass)) {
                 $schema->type = Generator::UNDEFINED;
                 $schema->ref = OA\Components::ref($typeSchema);
             }
         }
-    }
-
-    protected function augmentItems(OA\Schema $schema, Analysis $analysis): void
-    {
-        if (!Generator::isDefault($schema->type)) {
-            if (Generator::isDefault($schema->items)) {
-                $schema->items = new OA\Items([
-                    'type' => $schema->type,
-                    '_context' => new Context(['generated' => true], $schema->_context),
-                ]);
-
-                $this->type2ref($schema->items, $analysis);
-
-                $analysis->addAnnotation($schema->items, $schema->items->_context);
-
-                if (!Generator::isDefault($schema->ref)) {
-                    $schema->items->ref = $schema->ref;
-                    $schema->ref = Generator::UNDEFINED;
-                }
-            } elseif (Generator::isDefault($schema->items->type, $schema->items->oneOf, $schema->items->allOf, $schema->items->anyOf)) {
-                $schema->items->type = $schema->type;
-
-                $this->type2ref($schema->items, $analysis);
-            }
-        }
-
-        $this->mapNativeType($schema->items, $schema->items->type);
-        $schema->type = 'array';
     }
 
     /**
@@ -60,7 +31,7 @@ abstract class AbstractTypeResolver implements TypeResolverInterface
         if (is_array($type)) {
             $mapped = [];
             foreach ($type as $t) {
-                $mapped[] = $this->native2spec(strtolower($t));
+                $mapped[] = $this->native2spec(strtolower((string) $t));
             }
 
             $schema->type = $mapped;
