@@ -29,17 +29,17 @@ class CommandlineTest extends OpenApiTestCase
 
     public function testStdout(): void
     {
-        $basePath = $this->examplePath('petstore');
+        $basePath = static::examplePath('petstore');
         $path = "$basePath/annotations";
         exec($this->getCommandToExecute(__DIR__ . '/../bin/openapi --bootstrap ' . __DIR__ . '/cl_bootstrap.php --format yaml ' . escapeshellarg($path), '2>'), $output, $retval);
         $this->assertSame(0, $retval, implode(PHP_EOL, $output));
         $yaml = implode(PHP_EOL, $output);
-        $this->assertSpecEquals(file_get_contents($this->getSpecFilename('petstore')), $yaml);
+        $this->assertSpecEquals(file_get_contents(static::getSpecFilename('petstore')), $yaml);
     }
 
     public function testOutputToFile(): void
     {
-        $basePath = $this->examplePath('petstore');
+        $basePath = static::examplePath('petstore');
         $path = "$basePath/annotations";
         $filename = sys_get_temp_dir() . '/swagger-php-clitest.yaml';
         exec($this->getCommandToExecute(__DIR__ . '/../bin/openapi --bootstrap ' . __DIR__ . '/cl_bootstrap.php --format yaml -o ' . escapeshellarg($filename) . ' ' . escapeshellarg($path), '2>'), $output, $retval);
@@ -47,12 +47,12 @@ class CommandlineTest extends OpenApiTestCase
         $this->assertCount(0, $output, 'No output to stdout');
         $yaml = file_get_contents($filename);
         unlink($filename);
-        $this->assertSpecEquals(file_get_contents($this->getSpecFilename('petstore')), $yaml);
+        $this->assertSpecEquals(file_get_contents(static::getSpecFilename('petstore')), $yaml);
     }
 
     public function testAddProcessor(): void
     {
-        $basePath = $this->examplePath('petstore');
+        $basePath = static::examplePath('petstore');
         $path = "$basePath/annotations";
         $cmd = __DIR__ . '/../bin/openapi --bootstrap ' . __DIR__ . '/cl_bootstrap.php --processor OperationId --format yaml ' . escapeshellarg($path);
         exec($this->getCommandToExecute($cmd, '2>'), $output, $retval);
@@ -61,7 +61,7 @@ class CommandlineTest extends OpenApiTestCase
 
     public function testExcludeListWarning(): void
     {
-        $basePath = $this->examplePath('petstore');
+        $basePath = static::examplePath('petstore');
         $path = "$basePath/annotations";
         exec($this->getCommandToExecute(__DIR__ . '/../bin/openapi -e foo,bar ' . escapeshellarg($path) . ' 2>&1'), $output, $retval);
         $this->assertSame(1, $retval);
@@ -71,7 +71,7 @@ class CommandlineTest extends OpenApiTestCase
 
     public function testMissingArg(): void
     {
-        $basePath = $this->examplePath('petstore');
+        $basePath = static::examplePath('petstore');
         $path = "$basePath/annotations";
         exec($this->getCommandToExecute(__DIR__ . '/../bin/openapi ' . escapeshellarg($path) . ' -e 2>&1'), $output, $retval);
         $this->assertSame(1, $retval);
