@@ -135,55 +135,26 @@ class EncodingController
 
     }
 
-    #[OAT\Post(
-        path: '/endpoint/bc-encoding-attr',
-        requestBody: new OAT\RequestBody(
-            content: new OAT\MediaType(
-                mediaType: 'multipart/form-data',
-                schema: new OAT\Schema(
-                    properties: [
-                        new OAT\Property(property: 'encname', type: 'string'),
-                        ],
-                ),
-                encoding: [
-                    'encname' => [
-                       'contentType' => 'application/json',
-                    ],
-                ],
-            ),
-        ),
-        responses: [
-            new OAT\Response(
-                response: 200,
-                description: 'All good',
-            ),
-        ]
-    )]
-    public function bcEncodingAttr()
-    {
-
-    }
-
     /**
      * @OA\Post(
-     *     path="/endpoint/bc-encoding-annot",
+     *     path="/multipart-form-data-annot",
      *     @OA\RequestBody(
      *         @OA\MediaType(
      *             mediaType="multipart-form-data",
      *             @OA\Schema(
-     *                 @OA\Property(property="encname")
+     *                 @OA\Property(property="encname"),
+     *                 @OA\Property(
+     *                     property="other",
+     *                     @OA\Encoding(contentType="application/xml")
+     *                 )
      *             ),
-     *             encoding={
-     *                 "encname": {
-     *                     "contentType": "application/json"
-     *                 }
-     *             }
+     *             @OA\Encoding(property="encname", contentType="application/json")
      *         )
      *     ),
      *     @OA\Response(response="200", description="OK")
      * )
      */
-    public function bcEncodingAnnot()
+    public function multipartFormDataAnnot()
     {
 
     }
