@@ -6,7 +6,6 @@
 
 namespace OpenApi\Tests\Processors;
 
-use OpenApi\Analysers\ReflectionAnalyser;
 use OpenApi\Annotations as OA;
 use OpenApi\Generator;
 use OpenApi\Processors\AugmentProperties;
@@ -14,20 +13,20 @@ use OpenApi\Processors\AugmentSchemas;
 use OpenApi\Processors\MergeIntoComponents;
 use OpenApi\Processors\MergeIntoOpenApi;
 use OpenApi\Tests\OpenApiTestCase;
+use PHPUnit\Framework\Attributes\Group;
 
-/**
- * @group Properties
- */
+#[Group('Properties')]
 class AugmentPropertiesTest extends OpenApiTestCase
 {
     public function testAugmentProperties(): void
     {
-        $analysis = $this->analysisFromFixtures(['Customer.php']);
-        $analysis->process([
+        $analysis = $this->analysisFromFixtures([
+            'Customer.php',
+        ], $this->processorPipeline([
             new MergeIntoOpenApi(),
             new MergeIntoComponents(),
             new AugmentSchemas(),
-        ]);
+        ]));
 
         $customer = $analysis->openapi->components->schemas[0];
 
@@ -72,7 +71,7 @@ class AugmentPropertiesTest extends OpenApiTestCase
         $this->assertSame(Generator::UNDEFINED, $endorsedFriends->nullable);
         $this->assertSame(Generator::UNDEFINED, $endorsedFriends->allOf);
 
-        $analysis->process($this->initializeProcessors([new AugmentProperties()]));
+        $this->processorPipeline([new AugmentProperties()])->process($analysis);
 
         $expectedValues = [
             'property' => 'firstname',
@@ -150,16 +149,14 @@ class AugmentPropertiesTest extends OpenApiTestCase
 
     public function testTypedProperties(): void
     {
-        if ($this->getAnalyzer() instanceof ReflectionAnalyser && PHP_VERSION_ID < 70400) {
-            $this->markTestSkipped();
-        }
-
-        $analysis = $this->analysisFromFixtures(['TypedProperties.php']);
-        $analysis->process([
+        $analysis = $this->analysisFromFixtures([
+            'TypedProperties.php',
+        ], $this->processorPipeline([
             new MergeIntoOpenApi(),
             new MergeIntoComponents(),
             new AugmentSchemas(),
-        ]);
+        ]));
+
         [
             $stringType,
             $intType,
@@ -259,7 +256,7 @@ class AugmentPropertiesTest extends OpenApiTestCase
             'type' => Generator::UNDEFINED,
         ]);
 
-        $analysis->process($this->initializeProcessors([new AugmentProperties()]));
+        $this->processorPipeline([new AugmentProperties()])->process($analysis);
 
         $this->assertName($stringType, [
             'property' => 'stringType',
@@ -356,7 +353,7 @@ class AugmentPropertiesTest extends OpenApiTestCase
     protected function assertName(OA\Property $property, array $expectedValues): void
     {
         foreach ($expectedValues as $key => $val) {
-            $this->assertSame($val, $property->$key, '@OA\Property()->' . $key);
+            $this->assertSame($val, $property->$key, '@OA\Property()->property based on propertyname');
         }
     }
 }
