@@ -20,21 +20,22 @@ class AugmentRefsTest extends OpenApiTestCase
 
     public function testAugmentRefsForRequestBody(): void
     {
-        $analysis = $this->analysisFromFixtures(['Request.php']);
-        $analysis->process([
+        $analysis = $this->analysisFromFixtures([
+            'Request.php',
+        ], $this->processorPipeline([
             // create openapi->components
             new MergeIntoOpenApi(),
             // Merge standalone Scheme's into openapi->components
             new MergeIntoComponents(),
             new BuildPaths(),
             new AugmentRequestBody(),
-        ]);
+        ]));
 
         $this->assertSame($analysis->openapi->paths[0]->post->requestBody->ref, 'OpenApi\Tests\Fixtures\Request');
 
-        $analysis->process([
+        $this->processorPipeline([
             new AugmentRefs(),
-        ]);
+        ])->process($analysis);
 
         $this->assertSame($analysis->openapi->paths[0]->post->requestBody->ref, '#/components/requestBodies/Request');
     }
