@@ -16,18 +16,20 @@ class AugmentRequestBodyTest extends OpenApiTestCase
 {
     public function testAugmentSchemas(): void
     {
-        $analysis = $this->analysisFromFixtures(['Request.php']);
-        $analysis->process([
+        $analysis = $this->analysisFromFixtures([
+            'Request.php',
+        ], $this->processorPipeline([
             // create openapi->components
             new MergeIntoOpenApi(),
             // Merge standalone Scheme's into openapi->components
             new MergeIntoComponents(),
-        ]);
+        ]));
 
         $this->assertCount(1, $analysis->openapi->components->requestBodies);
         $request = $analysis->openapi->components->requestBodies[0];
         $this->assertSame(Generator::UNDEFINED, $request->request, 'Sanity check. No request was defined');
-        $analysis->process([new AugmentRequestBody()]);
+
+        $this->processorPipeline([new AugmentRequestBody()])->process($analysis);
 
         $this->assertSame('Request', $request->request, '@OA\RequestBody()->request based on classname');
     }
