@@ -24,7 +24,7 @@ class BuildPathsTest extends OpenApiTestCase
         ];
         $analysis = new Analysis([$openapi], $this->getContext());
         $analysis->openapi = $openapi;
-        $analysis->process([new BuildPaths()]);
+        $this->processorPipeline([new BuildPaths()])->process($analysis);
 
         $this->assertCount(1, $openapi->paths);
         $this->assertSame('/comments', $openapi->paths[0]->path);
@@ -41,10 +41,11 @@ class BuildPathsTest extends OpenApiTestCase
             ],
             $this->getContext()
         );
-        $analysis->process([
+        $this->processorPipeline([
             new MergeIntoOpenApi(),
             new BuildPaths(),
-        ]);
+        ])->process($analysis);
+
         $this->assertCount(1, $openapi->paths);
         $path = $openapi->paths[0];
         $this->assertSame('/comments', $path->path);
