@@ -27,7 +27,8 @@ class MergeIntoOpenApiTest extends OpenApiTestCase
         );
         $this->assertSame($openapi, $analysis->openapi);
         $this->assertSame(Generator::UNDEFINED, $openapi->info);
-        $analysis->process([new MergeIntoOpenApi()]);
+
+        $this->processorPipeline([new MergeIntoOpenApi()])->process($analysis);
 
         $this->assertSame($openapi, $analysis->openapi);
         $this->assertSame($info, $openapi->info);
