@@ -25,11 +25,12 @@ class MergeJsonContentTest extends OpenApiTestCase
 END;
         $analysis = new Analysis($this->annotationsFromDocBlockParser($comment), $this->getContext());
         $this->assertCount(3, $analysis->annotations);
-        /** @var OA\Response $response */
+
         $response = $analysis->getAnnotationsOfType(OA\Response::class)[0];
         $this->assertSame(Generator::UNDEFINED, $response->content);
         $this->assertCount(1, $response->_unmerged);
-        $analysis->process([new MergeJsonContent()]);
+
+        $this->processorPipeline([new MergeJsonContent()])->process($analysis);
 
         $this->assertIsArray($response->content);
         $this->assertCount(1, (array) $response->content);
@@ -49,10 +50,10 @@ END;
             )
 END;
         $analysis = new Analysis($this->annotationsFromDocBlockParser($comment), $this->getContext());
-        /** @var OA\Response $response */
         $response = $analysis->getAnnotationsOfType(OA\Response::class)[0];
         $this->assertCount(1, $response->content);
-        $analysis->process([new MergeJsonContent()]);
+
+        $this->processorPipeline([new MergeJsonContent()])->process($analysis);
 
         $this->assertCount(2, $response->content);
     }
@@ -67,12 +68,13 @@ END;
 END;
         $analysis = new Analysis($this->annotationsFromDocBlockParser($comment), $this->getContext());
         $this->assertCount(4, $analysis->annotations);
-        /** @var OA\Parameter $parameter */
+
         $parameter = $analysis->getAnnotationsOfType(OA\Parameter::class)[0];
         $this->assertSame(Generator::UNDEFINED, $parameter->content);
         $this->assertIsArray($parameter->_unmerged);
         $this->assertCount(1, $parameter->_unmerged);
-        $analysis->process([new MergeJsonContent()]);
+
+        $this->processorPipeline([new MergeJsonContent()])->process($analysis);
 
         $this->assertIsArray($parameter->content);
         $this->assertCount(1, (array) $parameter->content);
@@ -92,7 +94,7 @@ END;
             )
 END;
         $analysis = new Analysis($this->annotationsFromDocBlockParser($comment), $this->getContext());
-        $analysis->process([new MergeJsonContent()]);
+        $this->processorPipeline([new MergeJsonContent()])->process($analysis);
     }
 
     public function testInvalidParent(): void
@@ -106,6 +108,7 @@ END;
             )
 END;
         $analysis = new Analysis($this->annotationsFromDocBlockParser($comment), $this->getContext());
-        $analysis->process([new MergeJsonContent()]);
+        $this->processorPipeline([new MergeJsonContent()])->process($analysis);
+        ;
     }
 }
