@@ -7,20 +7,22 @@
 namespace OpenApi\Tests\Processors;
 
 use OpenApi\Annotations as OA;
-use OpenApi\Processors\OperationId;
 use OpenApi\Tests\OpenApiTestCase;
 
 class OperationIdTest extends OpenApiTestCase
 {
     public function testGeneratedOperationId(): void
     {
-        $analysis = $this->analysisFromFixtures([
-            'Processors/EntityControllerClass.php',
-            'Processors/EntityControllerInterface.php',
-            'Processors/EntityControllerTrait.php',
-        ]);
-        $analysis->process([new OperationId(false)]);
-        /** @var OA\Operation[] $operations */
+        $analysis = $this->analysisFromFixtures(
+            [
+                'Processors/EntityControllerClass.php',
+                'Processors/EntityControllerInterface.php',
+                'Processors/EntityControllerTrait.php',
+            ],
+            $this->processorPipeline(),
+            config: ['operationId' => ['hash' => false]]
+        );
+
         $operations = $analysis->getAnnotationsOfType(OA\Operation::class);
 
         $this->assertCount(3, $operations);
